+++ vercel.json
@@ -1,0 +1,29 @@
+{
+    "version": 2,
+    "builds": [
+      {
+        "src": "api/*.py",
+        "use": "@vercel/python"
+      },
+      {
+        "src": "public/**",
+        "use": "@vercel/static"
+      }
+    ],
+    "routes": [
+      {
+        "src": "/webhook",
+        "dest": "/api/webhook.py",
+        "methods": ["POST"]
+      },
+      {
+        "src": "/api/events",
+        "dest": "/api/webhook.py",
+        "methods": ["GET"]
+      },
+      {
+        "src": "/(.*)",
+        "dest": "/public/index.html"
+      }
+    ]
+  }
